#105 fix: udpate logic code
